Create ChamberDecompositionPolytopes
Compute the general chamber decomposition and each respective polytope for a given quiver. Polytopes are given as a list of vertices in the form Pn = (p1, p2, ..., pk) for each vertex n in the kth-dimensional Euclidean space. Each chamber of the chamber decomposition has a polytope associated with it, but different chambers can have the same polytope. The script outputs each chamber, the θ-stable spanning trees it is constructed by, and the vertices of the polytope it creates. Most importantly, the script will output the number of distinct polytopes.

Knowledge of the chamber decomposition is not necessary to make use of the polytope output, which is the primary use of the script. If information regarding the chamber decomposition is desired, the script outputs the dimension of the image of the I_Q map (which is a polytope), along with a list of facets and the primitive arrows (which describe this polytope's vertices) that define each facet.

This script utilizes three Macaulay2 packages, especially ThinSincereQuivers.